Extract DashboardDockManager, ModelInfoDialog, and move importReferenceData (#970)
Three more responsibilities extracted from ModelWindow:

- DashboardDockManager: owns pop-out/dock lifecycle for the dashboard
  panel, replacing 3 methods and 2 fields on ModelWindow
- ModelInfoDialog: self-contained dialog for editing model metadata
  (name, comment, author, source, license, URL)
- FileController.importReferenceData: reference data CSV import moved
  to FileController where other file I/O already lives

ModelWindow shrinks from 1222 to 1029 lines (-16%).
Combined with the previous commit: 1499 → 1029 (-31%).
